Moved absolute to relative conversion to the superstructure controller
