+++ src/main/java/frc/robot/subsystems/superstructure/arm/Arm.java
@@ -44,73 +44,8 @@ public class Arm extends GenericSuperstructure<Arm.ArmTarget> implements Loggabl
   /** The parent LoggableMechanism3d, typically a reference to the elevator subsystem */
   public LoggableMechanism3d loggableMechanism3dParent = null;
 
-  /**
-   * The target angle of the arm relative to the horizontal plane, changed later based on the
-   * current position
-   */
-  private Angle absoluteTargetAngleManual;
-
-  public void setAbsoluteTargetAngleManual(Angle angle) {
-    this.absoluteTargetAngleManual = angle;
-  }
-
-  public Angle getAbsoluteTargetAngleManual() {
-    return absoluteTargetAngleManual;
-  }
-
-  private ArmDirection armDirection = ArmDirection.BOTH;
-
-  public void setArmDirection(ArmDirection direction) {
-    this.armDirection = direction;
-  }
-
-  public ArmDirection getArmDirection() {
-    return armDirection;
-  }
-
-  /**
-   * Gets the relative angle to give to the motor controller to reach the given absolute angle
-   *
-   * @param absoluteAngle (0-360 degrees) centered at directly right when looking from the intake
-   *     side
-   * @return the relative angle to give to the motor controller
-   */
-  public double absoluteToRelativeTarget(double absoluteAngle) {
-    double currentAngle = getPosition();
-    double absolutePosition = currentAngle % 360.0;
-    double deltaAngle = normalizeAngle(absoluteAngle - absolutePosition);
-    double clockwise = absolutePosition + deltaAngle;
-    double counterClockwise = absolutePosition + deltaAngle - 360.0;
-
-    double finalTarget =
-        switch (armDirection) {
-          case CLOCKWISE -> clockwise;
-          case COUNTERCLOCKWISE -> counterClockwise;
-          case BOTH -> (Math.abs(deltaAngle) < Math.abs(deltaAngle - 360.0))
-              ? clockwise
-              : counterClockwise;
-        };
-    return finalTarget;
-  }
-
-  /**
-   * Normalizes between -360 and 360
-   *
-   * @param angle
-   * @return
-   */
-  public double normalizeAngle(double angle) {
-    double sign = Math.signum(angle);
-    angle = Math.abs(angle);
-    angle = angle % 360.0;
-    return angle * sign;
-  }
-
   @Override
   public void periodic() {
-    double relativeAngle = absoluteToRelativeTarget(absoluteTargetAngleManual.in(Units.Degrees));
-    super.setPositionTargetManual(relativeAngle);
-
     super.periodic();
 
     Logger.recordOutput(
